requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
clip
cmake==3.27.7
distro==1.8.0
exceptiongroup==1.1.3
filelock==3.13.1
fsspec==2023.10.0
icecream
gdown==4.7.1
google
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.19.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
lit==17.0.5
littleutils==0.2.2
MarkupSafe==2.1.3
modelscope
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
ogb==1.3.6
openai
opencv-python
outdated==0.2.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf
pydantic==1.10.13
pydantic_core==2.14.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex
requests==2.31.0
safetensors
scikit-learn
scipy
six
sniffio
soupsieve
sympy
threadpoolctl
tiktoken
tokenizers
torch
torchvision
tqdm
transformers
transformers_stream_generator
triton
typing_extensions
tzdata
urllib3
wilds
xformers
